fix: added db seeding
