+++ backend/api/index.ts
@@ -9,16 +9,13 @@ import MongoStore from "connect-mongo";
 import eventsRouter from "./routes/events";
 import authRouter from "./routes/auth";
 import ad_memRouter from "./routes/ad-mems";
+import { addTestMember, addTestEvents } from "./initTestData";
+import { Members } from "./models/Members";
 
 // SERVER DEFINITIONS
 const app = express();
 
 app.use(express.json({ limit: 300 }));
-
-// DATABASE CONNECTION
-mongoose.connect(process.env.MONGODB_URI!).then(() => {
-  console.log("Connected to DB");
-});
 
 // SESSION SETUP
 app.use(
@@ -36,16 +33,35 @@ app.use(
   })
 );
 
-// if (app.get("env") === "production") {
-//   app.set("trust proxy", 1); // trust first proxy
-// }
-
 app.use("/events", eventsRouter);
 app.use("/auth", authRouter);
 app.use("/members", ad_memRouter);
 
-app.listen(process.env.PORT, () => {
-  console.log(`Server started at ${process.env.PORT}`);
-});
+async function initializeApp() {
+  try {
+    // Connect to database
+    await mongoose.connect(process.env.MONGODB_URI!);
+    console.log("Connected to DB");
+
+    const adminExists: object | null = await Members.exists({
+      _id: process.env.ADMIN_EMAIL!
+    });
+
+    if (adminExists === null) {
+      await addTestMember();
+      await addTestEvents();
+    }
+
+    // Start server
+    app.listen(process.env.PORT, () => {
+      console.log(`Server started at ${process.env.PORT}`);
+    });
+  } catch (error) {
+    console.error("Failed to initialize app:", error);
+    process.exit(1);
+  }
+}
+
+initializeApp();
 
 export default app;
